sspi: add support of the security channel bindings for the Kerberos; small refactoring
