+++ src/sspi/ntlm/messages/computations.rs
@@ -6,11 +6,10 @@ use std::io::{self, Read, Write};
 use byteorder::{LittleEndian, ReadBytesExt, WriteBytesExt};
 use chrono::{DateTime, TimeZone, Utc};
 use lazy_static::lazy_static;
-use md5::{Digest, Md5};
 use rand::rngs::OsRng;
 use rand::Rng;
 
-use crate::crypto::{compute_hmac_md5, compute_md4, compute_md5, HASH_SIZE};
+use crate::crypto::{compute_hmac_md5, compute_md4, compute_md5, compute_md5_channel_bindings_hash, HASH_SIZE};
 use crate::ntlm::ChannelBindings;
 use crate::sspi::ntlm::messages::av_pair::*;
 use crate::sspi::ntlm::{
@@ -80,24 +79,6 @@ pub fn get_challenge_target_info(timestamp: u64) -> sspi::Result<Vec<u8>> {
     Ok(AvPair::list_to_buffer(&av_pairs)?)
 }
 
-pub fn calculate_channel_bindings_hash(channel_bindings: &ChannelBindings) -> [u8; HASH_SIZE] {
-    let mut context = Md5::new();
-    let mut result = [0x00; HASH_SIZE];
-
-    context.update(&channel_bindings.initiator_addr_type.to_be_bytes());
-    context.update(&channel_bindings.initiator.len().to_be_bytes());
-
-    context.update(&channel_bindings.acceptor_addr_type.to_be_bytes());
-    context.update(&channel_bindings.acceptor.len().to_be_bytes());
-
-    context.update(&channel_bindings.application_data.len().to_be_bytes());
-    context.update(&channel_bindings.application_data);
-
-    result.clone_from_slice(&context.finalize());
-
-    result
-}
-
 pub fn get_authenticate_target_info(
     target_info: &[u8],
     channel_bindings: &Option<ChannelBindings>,
@@ -121,7 +102,7 @@ pub fn get_authenticate_target_info(
     // because it is not used anywhere
 
     if let Some(channel_bindings) = channel_bindings {
-        av_pairs.push(AvPair::ChannelBindings(calculate_channel_bindings_hash(
+        av_pairs.push(AvPair::ChannelBindings(compute_md5_channel_bindings_hash(
             channel_bindings,
         )));
     }
